Login was refactorized and FormLogin was implemented
